src/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiomysql==0.2.0
aioodbc==0.5.0
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appscript==1.3.0
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bson==0.5.10
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect==1.2.0
Django==5.2.3
djangorestframework==3.16.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
geojson==2.5.0
gitdb==4.0.12
GitPython==3.1.44
google-trends-api==2.2.6
google_search_results==2.4.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imagekitio==4.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpickle==0.9.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
loguru==0.7.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
msgpack==1.1.0
multidict==6.1.0
mysql-connector-python==9.2.0
mysql-connector-python-rf==2.2.2
narwhals==1.27.1
numpy==2.2.3
openpyxl==3.1.5
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
passlib==1.7.4
paypalrestsdk==1.13.3
pillow==11.1.0
playwright==1.50.0
plotly==6.0.0
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pymongo==4.12.0
PyMySQL==1.1.1
pyngrok==7.2.8
pyodbc==5.2.0
pyOpenSSL==25.0.0
pyowm==3.3.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-weather==2.0.7
pytrends==4.9.2
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==0.10.1
rich==14.0.0
rich-toolkit==0.14.3
rpds-py==0.23.0
rsa==4.9.1
scraper==0.1.0
selenium==4.28.1
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.39
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.42.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
trio==0.29.0
trio-websocket==0.12.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.2
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.0.5
weather-api==1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xlwings==0.33.11
yarl==1.18.3
